Integrates weave.ts into error.ts
