+++ packages/sil-governance/src/routes/dashboard/components/Proposal/VotingPanel/Buttons.tsx
@@ -3,7 +3,7 @@ import Button from "@material-ui/core/Button";
 import { Block, Form, ToastContext, ToastVariant, useForm } from "medulas-react-components";
 import React, { useState } from "react";
 import * as ReactRedux from "react-redux";
-import { Weave } from "ui-logic";
+import { ErrorParser } from "ui-logic";
 
 import { communicationTexts } from "../../../../../communication";
 import { sendSignAndPostRequest } from "../../../../../communication/signandpost";
@@ -56,7 +56,7 @@ const Buttons = ({ id, vote }: Props): JSX.Element => {
       }
     } catch (error) {
       console.error(error);
-      const message = Weave.tryParseError(error) || "An unknown error occurred";
+      const message = ErrorParser.tryParseWeaveError(error) || "An unknown error occurred";
       toast.show(message, ToastVariant.ERROR);
     }
   };
